queue: Add redis_version and connection_version
This variables can be used to enable features within the code based on
availability in the running Redis server.

`connection_version` is the Redis server version.
`redis_version` is the `redis-py` Python library version.

As `distutils` is part of the Python stdlib it is used instead of more
recent external packages like `packaging`.

Signed-off-by: Paul Spooren <mail@aparcar.org>